split manager into three different use cases
