+++ services/core-api/app/api/verifiable_credentials/namespace.py
