+++ services/core-api/app/api/verifiable_credentials/untp_manager.py
@@ -138,39 +138,6 @@ def revoke_all_credentials_for_permit(permit_guid: str, mine_guid: str, reason: 
 
 
 @celery.task()
-def offer_newest_amendment_to_current_permittee(permit_amendment_guid: str,
-                                                cred_type: str = Config.CRED_DEF_ID_MINES_ACT_PERMIT
-                                                ):
-    """Revoke the existing credential and offer a new one with the newest amendment."""
-    newest_amendment = PermitAmendment.find_by_permit_amendment_guid(permit_amendment_guid)
-
-    permit = Permit.find_by_permit_guid(newest_amendment.permit_guid)
-    if permit.current_permittee_digital_wallet_connection_state != "active":
-        return "Permittee's wallet connection is not active, do not issue credential."
-
-    connection = PartyVerifiableCredentialConnection.find_active_by_party_guid(
-        permit.current_permittee_guid)
-
-    attributes = VerifiableCredentialManager.collect_attributes_for_mines_act_permit_111(
-        newest_amendment)
-
-    traction_svc = TractionService()
-    response = traction_svc.offer_mines_act_permit_111(connection.connection_id, attributes)
-    map_vc = PartyVerifiableCredentialMinesActPermit(
-        cred_exch_id=response["credential_exchange_id"],
-        cred_type=cred_type,
-        party_guid=permit.current_permittee_guid,
-        permit_amendment_guid=newest_amendment.permit_amendment_guid)
-
-    map_vc.save()
-
-    info_str = f"offer new_cred_exchange{response['credential_exchange_id']} for permit_amendment_guid={newest_amendment.permit_amendment_guid}"
-    current_app.logger.warning(info_str)         # not sure where to find this.
-
-    return info_str
-
-
-@celery.task()
 def process_all_untp_map_for_orgbook():
     """Find all permit amendments connected to orgbook verified parties, preprocess and sign any new credentials."""
 
@@ -385,7 +352,7 @@ def push_untp_map_data_to_publisher():
     return f"counts, published={success_count}, not_created={not_created_count}, skipped={skipped_count}, failed={len(failed_credentials)}"
 
 
-class VerifiableCredentialManager():
+class UNTPCredentialManager():
 
     def __init__(self):
         pass
@@ -472,243 +439,3 @@ class VerifiableCredentialManager():
             db.session.commit()
 
         return delete_count
-
-    @classmethod
-    def collect_attributes_for_mines_act_permit_111(
-            cls, permit_amendment: PermitAmendment) -> List[dict]:
-        # collect information for schema
-        # https://github.com/bcgov/bc-vcpedia/blob/main/credentials/bc-mines-act-permit/1.1.1/governance.md#261-schema-definition
-        credential_attrs = {}
-
-        #mine_types should related to the permits
-        #all mine_types for all permits are used in mine.mine_type
-        #but we only want the mine_type for this specific permit/permit_amendment.
-
-        #not really a 'mine_type' if it's managed at the permit level.
-        mine_types = [
-            mt for mt in permit_amendment.permit.site_properties
-            if mt.mine_guid == permit_amendment.mine_guid
-        ] if permit_amendment.permit.site_properties else None
-
-        #there my be mine types on other mines, so check again after filtering.
-        mine_type = mine_types[0] if mine_types else None
-
-        #provide permit object the permit_amendment mine_guid
-        permit_amendment.permit._context_mine = permit_amendment.mine_guid
-
-        mine_disturbance_list = []
-        mine_commodity_list = []
-
-        if mine_type:
-            mine_disturbance_list = [
-                mtd.mine_disturbance_literal for mtd in mine_type.mine_type_detail
-                if mtd.mine_disturbance_code
-            ]
-
-            mine_commodity_list = [
-                mtd.mine_commodity_literal for mtd in mine_type.mine_type_detail
-                if mtd.mine_commodity_code
-            ]
-
-        mine_status_xref = permit_amendment.mine.mine_status[0].mine_status_xref
-
-        credential_attrs["permit_no"] = permit_amendment.permit_no
-        credential_attrs["permit_status"] = permit_amendment.permit.permit_status_code_description
-        credential_attrs["permittee_name"] = permit_amendment.permit.current_permittee
-        credential_attrs[
-            "mine_operation_status"] = mine_status_xref.mine_operation_status.description
-        credential_attrs[
-            "mine_operation_status_reason"] = mine_status_xref.mine_operation_status_reason.description if mine_status_xref.mine_operation_status_reason else None
-        credential_attrs[
-            "mine_operation_status_sub_reason"] = mine_status_xref.mine_operation_status_sub_reason.description if mine_status_xref.mine_operation_status_sub_reason else None
-        credential_attrs["mine_disturbance"] = ", ".join(
-            mine_disturbance_list) if mine_disturbance_list else None
-        credential_attrs["mine_commodity"] = ", ".join(
-            mine_commodity_list) if mine_commodity_list else None
-        credential_attrs["mine_no"] = permit_amendment.mine.mine_no
-        credential_attrs["issue_date"] = int(permit_amendment.issue_date.strftime("%Y%m%d"))
-        # https://github.com/hyperledger/aries-rfcs/tree/main/concepts/0441-present-proof-best-practices#dates-and-predicates
-        credential_attrs["latitude"] = permit_amendment.mine.latitude
-        credential_attrs["longitude"] = permit_amendment.mine.longitude
-        credential_attrs["bond_total"] = permit_amendment.permit.active_bond_total
-        credential_attrs["tsf_operating_count"] = len([
-            tsf for tsf in permit_amendment.mine.mine_tailings_storage_facilities
-            if tsf.tsf_operating_status_code == "OPT"
-        ])
-        credential_attrs["tsf_care_and_maintenance_count"] = len([
-            tsf for tsf in permit_amendment.mine.mine_tailings_storage_facilities
-            if tsf.tsf_operating_status_code == "CAM"
-        ])
-
-        # offer credential
-        # "mime-type":"text/plain",
-        # NB Orbit does not expect this removing for now
-
-        attributes = [{
-            "name": str(attr),
-            "value": str(val),
-        } for attr, val in credential_attrs.items()]
-
-        return attributes
-
-    @classmethod
-    def produce_map_01_credential_payload(cls, did: str, permit_amendment: PermitAmendment):
-        #use w3c vcdm issue_date, not as an attribute in the credential
-        #https://www.w3.org/TR/vc-data-model/
-        id = permit_amendment.issue_date
-        #convert to datetime with tzinfo
-        issuance_date = datetime(id.year, id.month, id.day, 0, 0, 0, tzinfo=ZoneInfo("UTC"))
-        credential = {
-            "@context":
-            ["https://www.w3.org/2018/credentials/v1", {
-                "@vocab": "urn:bcgov:attributes#"
-            }],
-            "type": ["VerifiableCredential"],
-            "issuer": {
-                "id": did,
-            },
-            "issuanceDate": issuance_date.isoformat(),
-            "credentialSubject": {
-                a["name"]: a["value"]
-                for a in cls.collect_attributes_for_mines_act_permit_111(permit_amendment)
-                if a["name"] != "issue_date"
-            }
-        }
-        return credential
-
-    @classmethod
-    def produce_untp_cc_map_payload_without_id(cls, did: str,
-                                               permit_amendment: PermitAmendment) -> W3CCred | None:
-        """DEPRECATED: Produce payload for Mines Act Permit UNTP (v0.5.0) Conformity Credential from permit amendment and did."""
-
-        pmt_appts: List[MinePartyAppointment] = permit_amendment.permittee_appointments
-        current_app.logger.debug(
-            f"starting... produce_untp_cc_map_payload_without_id permit_amendment_guid={permit_amendment.permit_amendment_guid}"
-        )
-        permit_amendment_issue_date = permit_amendment.issue_date if isinstance(
-            permit_amendment.issue_date, date) else permit_amendment.issue_date.date()
-
-        if pmt_appts is None or len(pmt_appts) == 0:
-            current_app.logger.warning(
-                f"No permittee appointments found for permit_amendment_guid={permit_amendment.permit_amendment_guid}, cannot produce Mines Act Permit UNTP CC"
-            )
-            return None
-
-        #remove all appointments after the issue_date then take the top one, there are overlapping entries that may not be handled here.
-
-        try:
-            curr_appt = [
-                pa for pa in pmt_appts
-                if ensure_start_date_type(pa.start_date) <= permit_amendment_issue_date
-            ][0]
-        except IndexError:
-            current_app.logger.warning(
-                f"No valid permittee appointments found for permit_amendment_guid={permit_amendment.permit_amendment_guid}, cannot produce Mines Act Permit UNTP CC"
-            )
-            return None
-
-        orgbook_entity = curr_appt.party.party_bc_registration
-        if not orgbook_entity:
-            if curr_appt.party:
-                current_app.logger.warning(
-                    f"No Orgbook Entity for party_guid={curr_appt.party.party_guid}, could not produce Mines Act Permit UNTP CC"
-                )
-            else:
-                current_app.logger.error(
-                    f"No party for mine_party_appointment_id={curr_appt.mine_party_appt_id}, that shouldn't be possible"
-                )
-            return None
-
-        untp_party_cpo = base.Identifier(
-            id="did:web:untp.traceability.site:parties:regulators:CHIEF-PERMITTING-OFFICER",
-            name="Chief Permitting Officer of Mines",
-            registeredId=
-            "did:web:untp.traceability.site:parties:regulators:CHIEF-PERMITTING-OFFICER",
-            idScheme=base.IdentifierScheme(
-                id="https://w3c-ccg.github.io/did-method-web/", name="DID Web"))
-
-        business_registration_url = f"https://orgbook.gov.bc.ca/entity/{orgbook_entity.registration_id}"
-
-        untp_party_business = base.Party(
-            id=business_registration_url,
-            name=orgbook_entity.name_text,
-            registeredId=str(orgbook_entity.registration_id))
-
-        if not permit_amendment.mine.latitude or not permit_amendment.mine.longitude:
-            current_app.logger.warning(
-                f"Missing location information for permit_amendment_guid={permit_amendment.permit_amendment_guid}, cannot produce Mines Act Permit UNTP CC"
-            )
-            return None
-
-        facility = cc.Facility(
-            id=None,
-            name=permit_amendment.mine.mine_name,
-            registeredId=permit_amendment.mine.mine_no,
-            locationInformation=
-            f'https://plus.codes/{plus_code_encode(permit_amendment.mine.latitude, permit_amendment.mine.longitude)}',
-            address=None,
-            IDverifiedByCAB=True)
-
-        #TODO, can CORE identify commodities by their UNCEFACT code?
-        #remove duplicates
-        product_names = list(set([c for c in permit_amendment.mine.commodities]))
-        #sort list of strings for consistency
-        product_names.sort()
-
-        products = [cc.Product(id=None, name=c, IDverifiedByCAB=False) for c in product_names]
-
-        issue_date = permit_amendment.issue_date
-
-        untp_assessment = cc.ConformityAssessment(
-            id=None,
-            assessmentDate=issue_date,
-            referenceRegulation=cc.Regulation(
-                id="https://www.bclaws.gov.bc.ca/civix/document/id/complete/statreg/96293_01",
-                name="BC Mines Act",
-                jurisdictionCountry="CA",
-                administeredBy=base.Identifier(
-                    id="https://www2.gov.bc.ca/gov/content/home",
-                    name="Government of British Columbia",
-                    registeredId="BC-GOV",
-                    idScheme=base.IdentifierScheme(
-                        id="https://www2.gov.bc.ca/gov/content/home", name="BC-GOV")),
-                effectiveDate=datetime(2024, 5, 14, tzinfo=ZoneInfo("UTC")).isoformat()),
-            conformityTopic=codes.ConformityTopicCode.Governance_Compliance,
-            assessedFacility=[facility],
-            assessedProduct=products)
-
-        cred = UNTPCCMinesActPermit(
-            id=None,
-            name="Credential for permitNumber=" + permit_amendment.permit_no,
-            permitNumber=permit_amendment.permit_no,
-            assessmentLevel=codes.AssessmentLevelCode.GovtApproval,
-            attestationType=codes.AttestationType.Certification,
-            scope=cc.ConformityAssessmentScheme(
-                id=
-                "https://bcgov.github.io/digital-trust-toolkit/docs/governance/mining/bc-mines-act-permit/1.1.1/governance",
-                name="BC Mines Act Permit Credential (1.1.1) Governance Documentation"),
-            authorisation=[
-                base.Endorsement(
-                    id=
-                    "https://www2.gov.bc.ca/gov/content/industry/mineral-exploration-mining/permitting/mines-contact-info",
-                    name="BC Chief Permitting Officer of Mines",
-                    issuingAuthority=untp_party_cpo)
-            ],
-            issuedToParty=untp_party_business,
-            assessment=[untp_assessment])
-
-        w3c_cred = W3CCred(
-            id=None,                                                                                # to be populated after hashing.
-            type=[
-                "VerifiableCredential", "DigitalConformityCredential", "BCMinesActPermitCredential"
-            ],
-            issuer={"id": did},
-            validFrom=convert_date_to_iso_datetime(
-                permit_amendment.issue_date),                                                       #vcdm1.1, will change to 'validFrom' in vcdm2.0
-            credentialSubject=cred,
-            credentialSchema=[{
-                "id": Config.UNTP_DIGITAL_CONFORMITY_CREDENTIAL_CONTEXT,
-                "type": "JsonSchema"
-            }])
-
-        return w3c_cred
